+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All significant changes to this project will be documented here.
 
 ## [Unreleased]
 
+
+### Added
+
+* Specs now have a `replace` method that can be used to create a new instance
+  with some of the attributes replaced (similar to `namedtuple._replace`).
+
+
 ## [1.0]
 
 Release date: 2019-07-18